# docs/openapi.yaml
openapi: 3.0.3
info:
  title: SpeakSum API
  version: 2.0.0
  description: |
    SpeakSum 当前运行中的接口契约。

    当前主模型为：
    - content
    - summary_text
    - quotes
    - domains
    - domain graph
servers:
  - url: http://127.0.0.1:8000
    description: Local development
security:
  - BearerAuth: []
paths:
...
  /api/v1/contents:
    get:
      summary: List contents
      tags: [Contents]
      parameters:
        - name: q
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, processing, completed, ignored, failed, error]
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: page_size
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
      responses:
        '200':
          description: Content list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContentList'
  /api/v1/contents/{content_id}:
    get:
      summary: Get content detail
      tags: [Contents]
      parameters:
        - name: content_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Content detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Content'
        '404':
          description: Content not found
    delete:
      summary: Delete content
      tags: [Contents]
      parameters:
        - name: content_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Deleted
  /api/v1/contents/{content_id}/summary:
    patch:
      summary: Update content summary
      tags: [Contents]
      parameters:
        - name: content_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ContentSummaryUpdate'
      responses:
        '200':
          description: Updated content
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Content'
  /api/v1/contents/{content_id}/quotes/{quote_id}:
    patch:
      summary: Update quote
      tags: [Contents]
      parameters:
        - name: content_id
          in: path
          required: true
          schema:
            type: string
        - name: quote_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/QuoteUpdate'
      responses:
        '200':
          description: Updated quote
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Quote'
    delete:
      summary: Delete quote
      tags: [Contents]
      parameters:
        - name: content_id
          in: path
          required: true
          schema:
            type: string
        - name: quote_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Deleted
...
  /api/v1/knowledge-graph/domains/{domain_id}:
    get:
      summary: Get domain detail
      tags: [Knowledge Graph]
      parameters:
        - name: domain_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Domain detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GraphDomainDetail'
        '404':
          description: Domain not found
  /api/v1/knowledge-graph/layout:
    post:
      summary: Save graph layout
      tags: [Knowledge Graph]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GraphLayoutSaveRequest'
      responses:
        '200':
          description: Saved
  /api/v1/settings/model:
    get:
      summary: List model configs
      tags: [Settings]
      responses:
        '200':
          description: Model configs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ModelConfig'
    put:
      summary: Replace model configs
      tags: [Settings]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/ModelConfigCreate'
      responses:
        '200':
          description: Model configs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ModelConfig'
  /api/v1/settings/model/test:
    post:
      summary: Test model config
      tags: [Settings]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ModelConfigTestRequest'
      responses:
        '200':
          description: Test result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ModelConfigTestResponse'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Quote:
      type: object
      properties:
        id:
          type: string
        content_id:
          type: string
        sequence_number:
          type: integer
        text:
          type: string
        domain_ids:
          type: array
          items:
            type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
          nullable: true
    Content:
      type: object
      properties:
        id:
          type: string
        user_id:
          type: string
        title:
          type: string
        source_type:
          type: string
          enum: [meeting_minutes, other_text]
        content_date:
          type: string
          format: date
          nullable: true
        source_file_name:
          type: string
          nullable: true
        source_file_path:
          type: string
          nullable: true
        source_file_size:
          type: integer
          nullable: true
        file_type:
          type: string
          nullable: true
        status:
          type: string
        ignored_reason:
          type: string
          nullable: true
        error_message:
          type: string
          nullable: true
        summary_text:
          type: string
          nullable: true
        quotes:
          type: array
          items:
            $ref: '#/components/schemas/Quote'
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        completed_at:
          type: string
          format: date-time
          nullable: true
    ContentList:
      type: object
      properties:
        total:
          type: integer
        page:
          type: integer
        page_size:
          type: integer
        total_pages:
          type: integer
        items:
          type: array
          items:
            $ref: '#/components/schemas/Content'
    ContentSummaryUpdate:
      type: object
      required: [summary_text]
      properties:
        summary_text:
          type: string
    QuoteUpdate:
      type: object
      properties:
        text:
          type: string
        domain_ids:
          type: array
          items:
            type: string
    Domain:
      type: object
      properties:
        id:
          type: string
        display_name:
          type: string
        description:
          type: string
          nullable: true
        is_system_default:
          type: boolean
        sort_order:
          type: integer
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    KnowledgeGraphNode:
      type: object
      properties:
        id:
          type: string
        type:
          type: string
          example: domain
        label:
          type: string
        x:
          type: number
        y:
          type: number
        size:
          type: number
          nullable: true
    KnowledgeGraphEdge:
      type: object
      properties:
        source:
          type: string
        target:
          type: string
        type:
          type: string
        strength:
          type: number
          nullable: true
...
    GraphDomainQuote:
      type: object
      properties:
        id:
          type: string
        content_id:
          type: string
        text:
          type: string
        domain_ids:
          type: array
          items:
            type: string
    GraphDomainDetail:
      type: object
      properties:
        domain:
          $ref: '#/components/schemas/Domain'
        quotes:
          type: array
          items:
            $ref: '#/components/schemas/GraphDomainQuote'
        total:
          type: integer
    GraphLayoutSaveRequest:
      type: object
      properties:
        nodes:
          type: array
          items:
            $ref: '#/components/schemas/KnowledgeGraphNode'
        edges:
          type: array
          items:
            $ref: '#/components/schemas/KnowledgeGraphEdge'
        viewport:
          type: object
          nullable: true
    ModelConfigCreate:
      type: object
      required: [provider, name, default_model]
      properties:
        provider:
          type: string
        name:
          type: string
        api_key:
          type: string
          nullable: true
        base_url:
          type: string
          nullable: true
        default_model:
          type: string
        is_default:
          type: boolean
        is_enabled:
          type: boolean
    ModelConfig:
      type: object
      properties:
        id:
          type: string
        user_id:
          type: string
        provider:
          type: string
        name:
          type: string
        has_api_key:
          type: boolean
        base_url:
          type: string
          nullable: true
        default_model:
          type: string
        is_default:
          type: boolean
        is_enabled:
          type: boolean
        encryption_version:
          type: integer
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    ModelConfigTestRequest:
      type: object
      required: [provider, default_model]
      properties:
        config_id:
          type: string
          nullable: true
        provider:
          type: string
        api_key:
          type: string
          nullable: true
        base_url:
          type: string
          nullable: true
        default_model:
          type: string
    ModelConfigTestResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        provider:
          type: string
        model:
          type: string
...